+++ RELEASE_NOTES.md
@@ -11,7 +11,7 @@ Release Notes
 **Breaking changes**
 
 - `RowModel.edited` has been renamed `RowModel.databaseEdited`.
-- `RowModel.databaseTable` has been replaced with `RowModel.databaseTableName` which returns a String.
+- `RowModel.databaseTable` has been replaced with `RowModel.databaseTableName()` which returns a String.
 - `RowModel.setDatabaseValue(_:forColumn:)` has been removed. Use and override `RowModel.updateFromRow(_)` instead.